+++ src/main/java/quacky/QuackyException.java
@@ -1,6 +1,7 @@
-package Quacky;
+package quacky;
 
 import java.lang.Exception;
+
 public class QuackyException extends Exception {
     public QuackyException(String message) {
         super(message);
