+++ src/main/java/quacky/UI.java
@@ -1,6 +1,7 @@
-package Quacky;
+package quacky;
 
 import java.util.Scanner;
+
 public class UI {
     public boolean isRunning;
     private final Scanner scanner = new Scanner(System.in);
@@ -8,10 +9,12 @@ public class UI {
     public UI() {
         this.isRunning = true;
     }
+
     public String readCommand() {
         return scanner.nextLine();
     }
-    private String format(String text){
+
+    private String format(String text) {
         String[] lines = text.split("\n");
         StringBuilder sb = new StringBuilder("\t____________________________________________________________\n");
 
@@ -22,6 +25,7 @@ public class UI {
         sb.append("\t____________________________________________________________");
         return sb.toString();
     }
+
     public void say(String response) {
         System.out.println(format(response));
     }
@@ -29,16 +33,20 @@ public class UI {
     public void showGreeting() {
         this.say("Quack! how u doing, Im Quacky How can I help you?");
     }
+
     public void showFarewell() {
         this.isRunning = false;
         this.say("Quack Quack");
     }
+
     public void showList(TaskList tasks) {
         this.say(tasks.toString());
     }
+
     public void showMarkDone(String task) {
         this.say("Quack! I marked this task as done \n\t" + task);
     }
+
     public void showUnmarkDone(String task) {
         this.say("Quack! I marked this task as not done \n\t" + task);
     }
@@ -48,11 +56,13 @@ public class UI {
                 + "\nNow you have " + taskNumber + " tasks in the list.";
         this.say(message);
     }
+
     public void showAddTask(int taskNumber, String task) {
         String message = "Got it. I've added this task:\n\t" + task +
-                        "\nNow you have " + taskNumber + " tasks in the list.";
+                "\nNow you have " + taskNumber + " tasks in the list.";
         this.say(message);
     }
+
     void showErrorMessage(Exception e) {
         this.say(e.getMessage());
     }
