+++ src/test/java/quacky/DeadlineTest.java
@@ -1,4 +1,5 @@
-package Quacky;
+package quacky;
+
 import org.junit.jupiter.api.Test;
 
 import java.time.LocalDate;
