+++ text-ui-test/EXPECTED.TXT
@@ -1,5 +1,5 @@
 	____________________________________________________________
-	Quack! how u doing, Im Quacky.Quacky. How can I help you?
+	Quack! how u doing, Im quacky.quacky. How can I help you?
 	____________________________________________________________
 	____________________________________________________________
 	Quack! I've added this task:
